refactor view into cell
