--- CIE-DEMO/Components/Shared/Collections/SimilarMoviesCollectionView.swift
@@ -1,13 +1,0 @@
-//
-//  SimilarMoviesView.swift
-//  CIE-DEMO
-//
-//  Created by hassan uriostegui on 9/21/19.
-//  Copyright © 2019 eonflux. All rights reserved.
-//
-
-import UIKit
-
-class SimilarMoviesView: UIView {
-
-}
